+++ forui/lib/src/theme/colors.dart
@@ -68,6 +68,26 @@ final class FColors with Diagnosticable {
   /// Typically used to color widgets against a [background].
   final Color foreground;
 
+  /// The card color.
+  ///
+  /// Typically used as a background for card widgets.
+  final Color card;
+
+  /// The card foreground color.
+  ///
+  /// Typically used to color widgets against a [card] colored background.
+  final Color cardForeground;
+
+  /// The popover color.
+  ///
+  /// Typically used as a background for popover widgets such as tooltips and dropdowns.
+  final Color popover;
+
+  /// The popover foreground color.
+  ///
+  /// Typically used to color widgets against a [popover] colored background.
+  final Color popoverForeground;
+
   /// The primary color.
   ///
   /// Typically used as a background for [primaryForeground] colored widgets.
@@ -121,6 +141,11 @@ final class FColors with Diagnosticable {
   /// The border color.
   final Color border;
 
+  /// The field color.
+  ///
+  /// Typically used as a background for text fields.
+  final Color field;
+
   /// The percentage to lighten dark colors by. A higher value will result in a more pronounced lightening effect.
   ///
   /// Defaults to 0.075.
@@ -153,6 +178,10 @@ final class FColors with Diagnosticable {
     required this.barrier,
     required this.background,
     required this.foreground,
+    required this.card,
+    required this.cardForeground,
+    required this.popover,
+    required this.popoverForeground,
     required this.primary,
     required this.primaryForeground,
     required this.secondary,
@@ -164,6 +193,7 @@ final class FColors with Diagnosticable {
     required this.error,
     required this.errorForeground,
     required this.border,
+    required this.field,
     this.hoverLighten = 0.075,
     this.hoverDarken = 0.05,
     this.disabledOpacity = 0.5,
@@ -178,6 +208,10 @@ final class FColors with Diagnosticable {
     barrier: FColors.lerpColor(a.barrier, b.barrier, t)!,
     background: FColors.lerpColor(a.background, b.background, t)!,
     foreground: FColors.lerpColor(a.foreground, b.foreground, t)!,
+    card: FColors.lerpColor(a.card, b.card, t)!,
+    cardForeground: FColors.lerpColor(a.cardForeground, b.cardForeground, t)!,
+    popover: FColors.lerpColor(a.popover, b.popover, t)!,
+    popoverForeground: FColors.lerpColor(a.popoverForeground, b.popoverForeground, t)!,
     primary: FColors.lerpColor(a.primary, b.primary, t)!,
     primaryForeground: FColors.lerpColor(a.primaryForeground, b.primaryForeground, t)!,
     secondary: FColors.lerpColor(a.secondary, b.secondary, t)!,
@@ -189,6 +223,7 @@ final class FColors with Diagnosticable {
     error: FColors.lerpColor(a.error, b.error, t)!,
     errorForeground: FColors.lerpColor(a.errorForeground, b.errorForeground, t)!,
     border: FColors.lerpColor(a.border, b.border, t)!,
+    field: FColors.lerpColor(a.field, b.field, t)!,
     hoverLighten: lerpDouble(a.hoverLighten, b.hoverLighten, t)!,
     hoverDarken: lerpDouble(a.hoverDarken, b.hoverDarken, t)!,
     disabledOpacity: lerpDouble(a.disabledOpacity, b.disabledOpacity, t)!,
@@ -257,6 +292,10 @@ final class FColors with Diagnosticable {
     Color? barrier,
     Color? background,
     Color? foreground,
+    Color? card,
+    Color? cardForeground,
+    Color? popover,
+    Color? popoverForeground,
     Color? primary,
     Color? primaryForeground,
     Color? secondary,
@@ -268,6 +307,7 @@ final class FColors with Diagnosticable {
     Color? error,
     Color? errorForeground,
     Color? border,
+    Color? field,
     double? hoverLighten,
     double? hoverDarken,
     double? disabledOpacity,
@@ -277,6 +317,10 @@ final class FColors with Diagnosticable {
     barrier: barrier ?? this.barrier,
     background: background ?? this.background,
     foreground: foreground ?? this.foreground,
+    card: card ?? this.card,
+    cardForeground: cardForeground ?? this.cardForeground,
+    popover: popover ?? this.popover,
+    popoverForeground: popoverForeground ?? this.popoverForeground,
     primary: primary ?? this.primary,
     primaryForeground: primaryForeground ?? this.primaryForeground,
     secondary: secondary ?? this.secondary,
@@ -288,6 +332,7 @@ final class FColors with Diagnosticable {
     error: error ?? this.error,
     errorForeground: errorForeground ?? this.errorForeground,
     border: border ?? this.border,
+    field: field ?? this.field,
     hoverLighten: hoverLighten ?? this.hoverLighten,
     hoverDarken: hoverDarken ?? this.hoverDarken,
     disabledOpacity: disabledOpacity ?? this.disabledOpacity,
@@ -302,6 +347,10 @@ final class FColors with Diagnosticable {
       ..add(ColorProperty('barrier', barrier))
       ..add(ColorProperty('background', background))
       ..add(ColorProperty('foreground', foreground))
+      ..add(ColorProperty('card', card))
+      ..add(ColorProperty('cardForeground', cardForeground))
+      ..add(ColorProperty('popover', popover))
+      ..add(ColorProperty('popoverForeground', popoverForeground))
       ..add(ColorProperty('primary', primary))
       ..add(ColorProperty('primaryForeground', primaryForeground))
       ..add(ColorProperty('secondary', secondary))
@@ -313,6 +362,7 @@ final class FColors with Diagnosticable {
       ..add(ColorProperty('error', error))
       ..add(ColorProperty('errorForeground', errorForeground))
       ..add(ColorProperty('border', border))
+      ..add(ColorProperty('field', field))
       ..add(PercentProperty('hoverLighten', hoverLighten))
       ..add(PercentProperty('hoverDarken', hoverDarken))
       ..add(PercentProperty('disabledOpacity', disabledOpacity));
@@ -327,6 +377,10 @@ final class FColors with Diagnosticable {
           barrier == other.barrier &&
           background == other.background &&
           foreground == other.foreground &&
+          card == other.card &&
+          cardForeground == other.cardForeground &&
+          popover == other.popover &&
+          popoverForeground == other.popoverForeground &&
           primary == other.primary &&
           primaryForeground == other.primaryForeground &&
           secondary == other.secondary &&
@@ -338,6 +392,7 @@ final class FColors with Diagnosticable {
           error == other.error &&
           errorForeground == other.errorForeground &&
           border == other.border &&
+          field == other.field &&
           hoverLighten == other.hoverLighten &&
           hoverDarken == other.hoverDarken &&
           disabledOpacity == other.disabledOpacity;
@@ -349,6 +404,10 @@ final class FColors with Diagnosticable {
       barrier.hashCode ^
       background.hashCode ^
       foreground.hashCode ^
+      card.hashCode ^
+      cardForeground.hashCode ^
+      popover.hashCode ^
+      popoverForeground.hashCode ^
       primary.hashCode ^
       primaryForeground.hashCode ^
       secondary.hashCode ^
@@ -360,6 +419,7 @@ final class FColors with Diagnosticable {
       error.hashCode ^
       errorForeground.hashCode ^
       border.hashCode ^
+      field.hashCode ^
       hoverLighten.hashCode ^
       hoverDarken.hashCode ^
       disabledOpacity.hashCode;