+++ forui/lib/src/widgets/alert.dart
@@ -135,13 +135,13 @@ extension type FAlertStyles._(FVariants<FAlertVariantConstraint, FAlertStyle, FA
     : this._(
         .delta(
           FAlertStyle(
-            iconStyle: IconThemeData(color: colors.foreground, size: 20),
-            titleTextStyle: typography.base.copyWith(fontWeight: .w500, color: colors.foreground, height: 1.2),
-            subtitleTextStyle: typography.sm.copyWith(color: colors.foreground),
+            iconStyle: IconThemeData(color: colors.cardForeground, size: 20),
+            titleTextStyle: typography.base.copyWith(fontWeight: .w500, color: colors.cardForeground, height: 1.2),
+            subtitleTextStyle: typography.sm.copyWith(color: colors.mutedForeground),
             decoration: BoxDecoration(
               border: .all(color: colors.border),
               borderRadius: style.borderRadius,
-              color: colors.background,
+              color: colors.card,
             ),
           ),
           variants: {