+++ forui/test/src/theme/typography_test.dart
@@ -56,6 +56,10 @@ void main() {
         barrier: Colors.black12,
         background: Colors.black,
         foreground: Colors.black12,
+        card: Colors.cyan,
+        cardForeground: Colors.cyanAccent,
+        popover: Colors.deepOrange,
+        popoverForeground: Colors.deepOrangeAccent,
         primary: Colors.black26,
         primaryForeground: Colors.black38,
         secondary: Colors.black45,
@@ -67,6 +71,7 @@ void main() {
         error: Colors.red,
         errorForeground: Colors.redAccent,
         border: Colors.lightBlue,
+        field: Colors.lime,
       );
 
       test('no arguments', () {